cp2_cp4/requirements-pi.txt
paho-mqtt==2.1.0
pyserial==3.5
opencv-python==4.10.0.84
numpy==1.26.4
